refactor: seperate community list hook (#84)
