+++ src/hooks/api/community/useCommunityListQuery.ts
@@ -1,10 +1,23 @@
 import { type UseSuspenseQueryOptions, useSuspenseQuery } from "@tanstack/react-query";
 
-import { getCommunityList, getRecommendedCommunityList } from "@/apis/community/getCommunityList";
+import { GET, createInit } from "@/apis/httpMethod";
 
+import { END_POINTS } from "@/constants/api";
+import { INITIAL_FILTER_OPTION_DATA } from "@/constants/community";
 import { QUERY_KEYS } from "@/constants/queryKey";
 
 import type { FilterSelectedType, GetCommunityListResponseType } from "@/types/community";
+
+const getCommunityList = async (filterOption?: FilterSelectedType) => {
+  const options = filterOption ?? INITIAL_FILTER_OPTION_DATA;
+
+  const data = await GET<GetCommunityListResponseType>(
+    END_POINTS.COMMUNITY_LIST(options.search, options.tags, options.sort, options.date),
+    createInit(),
+  );
+
+  return data;
+};
 
 export const communityListQueryOptions = (
   filterOption?: FilterSelectedType,
@@ -13,16 +26,6 @@ export const communityListQueryOptions = (
   queryFn: async () => await getCommunityList(filterOption),
 });
 
-export const RecommendedCommunityListQueryOptions =
-  (): UseSuspenseQueryOptions<GetCommunityListResponseType> => ({
-    queryKey: [QUERY_KEYS.COMMUNITY_LIST_RECOMMEND],
-    queryFn: async () => await getRecommendedCommunityList(),
-  });
-
 export const useCommunityListQuery = (filterOption: FilterSelectedType) => {
   return useSuspenseQuery(communityListQueryOptions(filterOption));
-};
-
-export const useRandomCommunityListQuery = () => {
-  return useSuspenseQuery(RecommendedCommunityListQueryOptions());
 };
